Add KECCAK256 to jolt-inlines
