+++ Cargo.toml
@@ -24,6 +24,7 @@ keywords = ["SNARK", "cryptography", "proofs"]
 members = [
     "jolt-core",
     "jolt-inlines/sha2",
+    "jolt-inlines/keccak256",
     "tracer",
     "common",
     "jolt-sdk",
@@ -110,4 +111,4 @@ ark-bn254 = { git = "https://github.com/a16z/arkworks-algebra", branch = "dev/tw
 ignored = ["jolt-sdk"]
 
 [workspace.dependencies]
-spinners = { version= "4.1.1", default-features = false }
+spinners = { version = "4.1.1", default-features = false }
