+++ jolt-inlines/keccak256/src/lib.rs
@@ -1,0 +1,26 @@
+//! Keccak-256 inline.
+
+#![cfg_attr(not(feature = "host"), no_std)]
+
+pub const NUM_LANES: usize = 25;
+pub type Keccak256State = [u64; NUM_LANES];
+
+pub mod sdk;
+pub use sdk::*;
+
+#[cfg(feature = "host")]
+pub mod exec;
+#[cfg(feature = "host")]
+pub mod trace_generator;
+
+#[cfg(feature = "host")]
+mod host;
+#[cfg(feature = "host")]
+pub use host::*;
+
+#[cfg(all(test, feature = "host"))]
+pub mod test_constants;
+#[cfg(all(test, feature = "host"))]
+pub mod test_utils;
+#[cfg(all(test, feature = "host"))]
+mod tests;
